--- .streamlit/config.toml
@@ -1,3 +1,0 @@
-[theme]
-base="dark"
-backgroundColor="#000000"
