+++ DeepFeelings/.streamlit/config.toml
@@ -1,0 +1,6 @@
+[theme]
+primaryColor = "#E694FF"
+backgroundColor = "#00172B"
+secondaryBackgroundColor = "#0083B8"
+textColor = "#C6CDD4"
+font = "sans serif"
